+++ src/animation/offline/tools/import2ozz_config.h
@@ -32,7 +32,50 @@
 
 #include <json/json-forwards.h>
 
+namespace ozz {
+namespace animation {
+namespace offline {
+
 // Get the sanitized (all members are set, with the right types) configuration.
 bool ProcessConfiguration(Json::Value* _config);
+
+template <typename _Type, typename _Enum>
+struct JsonEnum {
+  // Struct allowing inheriting class to provide enum names.
+  struct EnumNames {
+    size_t count;
+    const char** names;
+  };
+
+  static bool GetEnumFromName(const char* _name, _Enum* _enum) {
+    const EnumNames enums = _Type::GetNames();
+    for (size_t i = 0; i < enums.count; ++i) {
+      if (strcmp(enums.names[i], _name) == 0) {
+        *_enum = static_cast<_Enum>(i);
+        return true;
+      }
+    }
+    return false;
+  }
+
+  static const char* GetEnumName(_Enum _enum) {
+    const EnumNames enums = _Type::GetNames();
+    assert(static_cast<size_t>(_enum) < enums.count);
+    return enums.names[_enum];
+  }
+
+  static bool IsValidEnumName(const char* _name) {
+    const EnumNames enums = _Type::GetNames();
+    bool valid = false;
+    for (size_t i = 0; !valid && i < enums.count; ++i) {
+      valid = strcmp(enums.names[i], _name) == 0;
+    }
+    return valid;
+  }
+};
+
+}  // namespace offline
+}  // namespace animation
+}  // namespace ozz
 
 #endif  // OZZ_ANIMATION_OFFLINE_TOOLS_CONFIGURATION_H_